+++ src/main.ts
@@ -1,18 +1,20 @@
 import { setFailed } from '@actions/core'
 import { Conf, PR } from './types'
-import getCardIds from './actions/getCardIds'
-import addCardsLinkToPullRequest from './actions/addCardsLinkToPullRequest'
-import addPullRequestLinkToCards from './actions/addPullRequestLinkToCards'
-import moveOrArchiveCards from './actions/moveOrArchiveCards'
-import addLabelToCards from './actions/addLabelToCards'
-import updateCardMembers from './actions/updateCardMembers'
+import {
+	addCardLinksToPullRequest,
+	addLabelToCards,
+	addPullRequestLinkToCards,
+	getCardIds,
+	moveOrArchiveCards,
+	updateCardMembers,
+} from './actions'
 
 export async function run(pr: PR, conf: Conf = {}) {
 	try {
 		const cardIds = await getCardIds(conf, pr)
 
 		if (cardIds.length) {
-			await addCardsLinkToPullRequest(conf, cardIds, pr)
+			await addCardLinksToPullRequest(conf, cardIds, pr)
 			await addPullRequestLinkToCards(cardIds, pr.html_url || pr.url)
 			await moveOrArchiveCards(conf, cardIds, pr)
 			await addLabelToCards(conf, cardIds, pr.head)
